+++ system/src/Grav/Framework/Flex/Interfaces/FlexStorageInterface.php
@@ -25,6 +25,11 @@ interface FlexStorageInterface
     public function __construct(array $options);
 
     /**
+     * @return string
+     */
+    public function getKeyField(): string;
+
+    /**
      * Returns associated array of all existing storage keys with a timestamp.
      *
      * @return  array Returns all existing keys as `[key => [storage_key => key, storage_timestamp => timestamp], ...]`.