+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     * Fixed deleting last list item in the form
     * Grav 1.6: Fixed Flex erroring out with all numeric keys in storage
     * Grav 1.6: Fixed error in `FlexUserIndex` if there are users without an email address
+    * Grav 1.6: Fixed bad key in Flex Users if file storage is being used
 
 # v1.6.0-rc.4
 ## 03/20/2019